Add packaged known_hosts file to UserKnownHosts.
This adds the packaged ssh_known_hosts file, if it exists, to the
end of the UserKnownHosts option.  This means that if there are no
other host keys for a given server, the packaged ones will be used
rather than throwing a (silent) host key error.  If an unlisted
server, or a listed server wth an existing host key of a different
type is used, the behaviour remains unchanged.